-Harmonize skillid, skill_num, skillnum, skill, to skill_id and same for skill_lv -Optimise type from int to int16 in order to reduce ram consumtion.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@17065 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/instance.h
@@ -22,7 +22,7 @@ struct s_instance {
 	int users;
 
 	struct DBMap* vars; // Instance Variable for scripts
-	
+
 	int progress_timer;
 	time_t progress_timeout;
 
@@ -35,9 +35,9 @@ extern struct s_instance instance[MAX_INSTANCE];
 
 int instance_create(int party_id, const char *name);
 int instance_add_map(const char *name, int instance_id, bool usebasename);
-void instance_del_map(int m);
-int instance_map2imap(int m, int instance_id);
-int instance_mapid2imapid(int m, int instance_id);
+void instance_del_map(int16 m);
+int instance_map2imap(int16 m, int instance_id);
+int instance_mapid2imapid(int16 m, int instance_id);
 void instance_destroy(int instance_id);
 void instance_init(int instance_id);
